+++ backend/config/settings/local.py
@@ -72,6 +72,8 @@ REST_FRAMEWORK["DEFAULT_AUTHENTICATION_CLASSES"].append(  # noqa: F405
 
 # Custom Storage configuration to reuse minio values
 # ------------------------------------------------------------------------------
+# Storage for browser access via signed URLs
+
 STORAGES["default"]["OPTIONS"].update(  # type: ignore[index]  # noqa: F405
     {
         "access_key": env.str("MINIO_ROOT_USER"),
@@ -83,6 +85,17 @@ STORAGES["default"]["OPTIONS"].update(  # type: ignore[index]  # noqa: F405
         "url_protocol": "http:",
     }
 )
+STORAGES["browser"] = {  # noqa: F405
+    "BACKEND": "utils.storage.BrowserS3Storage",
+    "OPTIONS": STORAGES["default"]["OPTIONS"],  # noqa: F405
+}
+
+# Storage for internal docker network access
+STORAGES["internal"] = {  # noqa: F405
+    "BACKEND": "utils.storage.InternalS3Storage",
+    "OPTIONS": STORAGES["default"]["OPTIONS"],  # noqa: F405
+}
+
 # Your stuff...
 # ------------------------------------------------------------------------------
 EMAIL_SUBJECT_PREFIX = "[coordnet.dev - DEV]"