+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 - Ability for Skills to write to Spaces (using External Nodes).
+- Add API support for creating and managing PaperQA collections.
 
 ## [25.5.2] - 2025-05-19